refactor: get rid of luxon in frontend
