+++ frontend/package.json
@@ -28,7 +28,6 @@
     "graphql": "^16.8.1",
     "graphql-tag": "^2.12.6",
     "lodash.debounce": "^4.0.8",
-    "luxon": "^3.4.3",
     "mark.js": "^8.11.1",
     "moment": "^2.29.4",
     "qs": "^6.11.2",
@@ -65,7 +64,6 @@
     "@types/gapi.auth2": "^0.0.60",
     "@types/gapi.calendar": "^3.0.9",
     "@types/lodash.debounce": "^4.0.9",
-    "@types/luxon": "^3.3.2",
     "@types/mark.js": "^8.11.9",
     "@types/node": "^20.8.2",
     "@types/qs": "^6.9.8",
